Menu and usability improvements: chg - Moved category management to a separate dialog that can be open by "Menu->Settings->Categories...". chg - Renamed "Menu->Settings->Edit..." to "Menu->Settings->Options...". chg - SettingsDialog: Added OK/Cancel buttons to provide the ability to dismiss changes instead of being forced to save. fix - Resolved #74: Column "Hidden" visibility wasn't updated properly. fix - "Menu->Settings->Show hidden mods" wasn't updated when changed in Options dialog.
